# docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Smart Attendance System API
  version: 1.0.0
  description: REST API for Smart Attendance System
servers:
  - url: http://localhost:5000/api
    description: Development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with Firebase ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken:
                  type: string
              required: [idToken]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      user:
                        $ref: '#/components/schemas/User'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Token refreshed

  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User profile

  /teacher/timetable:
    get:
      tags: [Teacher]
      summary: Get teacher timetable
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Timetable entries

  /teacher/current-class:
    get:
      tags: [Teacher]
      summary: Auto-detect current class
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current class info or null

  /teacher/attendance/start:
    post:
      tags: [Teacher]
      summary: Start attendance session
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                classId: { type: string }
                subject: { type: string }
      responses:
        '201':
          description: Session created

  /teacher/{classId}/attendance:
    post:
      tags: [Teacher]
      summary: Mark student attendance
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: classId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId: { type: string }
                status: { type: string, enum: [present, absent, late] }
      responses:
        '200':
          description: Attendance marked
    get:
      tags: [Teacher]
      summary: Get attendance history for class
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: classId
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: Paginated attendance records

  /teacher/{classId}/attendance/submit:
    post:
      tags: [Teacher]
      summary: Submit and finalize attendance
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: classId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Attendance submitted

  /teacher/{classId}/students:
    get:
      tags: [Teacher]
      summary: Get students in class
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: classId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Student list

  /student/dashboard:
    get:
      tags: [Student]
      summary: Get attendance dashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard with subject stats and alerts

  /student/{id}/attendance:
    get:
      tags: [Student]
      summary: Get student attendance history
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Attendance history

  /student/{id}/calendar:
    get:
      tags: [Student]
      summary: Calendar view of attendance
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: month
          schema: { type: integer }
        - in: query
          name: year
          schema: { type: integer }
      responses:
        '200':
          description: Calendar data

  /admin/users:
    get:
      tags: [Admin]
      summary: List users (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: role
          schema: { type: string, enum: [teacher, student, admin] }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: Paginated user list
    post:
      tags: [Admin]
      summary: Create user
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created

  /admin/users/{id}:
    put:
      tags: [Admin]
      summary: Update user
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User updated
    delete:
      tags: [Admin]
      summary: Delete user
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User deleted

  /admin/reports:
    get:
      tags: [Admin]
      summary: Get attendance reports
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: startDate
          schema: { type: string, format: date }
        - in: query
          name: endDate
          schema: { type: string, format: date }
      responses:
        '200':
          description: Reports with summary

  /admin/reports/export/excel:
    get:
      tags: [Admin]
      summary: Export as Excel
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Excel file download

  /admin/reports/export/pdf:
    get:
      tags: [Admin]
      summary: Export as PDF
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: PDF file download

  /admin/config:
    get:
      tags: [Admin]
      summary: Get system config
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Config object
    put:
      tags: [Admin]
      summary: Update system config
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Config updated

  /admin/logs:
    get:
      tags: [Admin]
      summary: Get system logs
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Paginated logs

  /sync/queue:
    post:
      tags: [Sync]
      summary: Process offline sync queue
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      action: { type: string }
                      data: { type: object }
                      createdAt: { type: string }
      responses:
        '200':
          description: Sync results

  /sync/status:
    get:
      tags: [Sync]
      summary: Get sync queue status
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Pending and failed counts

components:
  schemas:
    User:
      type: object
      properties:
        uid: { type: string }
        email: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [teacher, student, admin] }
        rollNumber: { type: string }
        studentId: { type: string }
        department: { type: string }
        photoURL: { type: string }

    CreateUser:
      type: object
      required: [email, password, displayName, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        displayName: { type: string }
        role: { type: string, enum: [teacher, student, admin] }
        rollNumber: { type: string }
        studentId: { type: string }
        department: { type: string }